Agency Payable Search - Insert
git-svn-id: http://nt-svn-jira.vdc.csc.com/HPF/Implementation/Source Code@616 6d9d970a-e046-46de-96ce-fbc9f6c46dd9

+++ HPF.FutureState/HPF.FutureState.Common/DataTransferObjects/AgencyPayableCaseDTO.cs
@@ -12,7 +12,7 @@ namespace HPF.FutureState.Common.DataTransferObjects
         public int ForeclosureCaseId { get; set; }
         public int AgencyPayableId { get; set; }
         public DateTime PaymentDate { get; set; }
-        public double PaymentAmount { get; set; }
+        public decimal PaymentAmount { get; set; }
         public string NFMCDiffererencePaidIndicator { get; set; }
     }
 }
